feat: add environment validation scripts
