+++ lib/validate-server.ts
@@ -1,4 +1,3 @@
-import { z } from 'zod';
 import type { ZodTypeAny } from 'zod';
 import crypto from 'crypto';
 import type { NextApiRequest, NextApiResponse } from 'next';
@@ -73,19 +72,4 @@ export function validateRequest(
   }
 
   return { query: parsedQuery, body: parsedBody };
-}
-
-const EnvSchema = z.object({
-  JWT_SECRET: z.string().min(1, 'JWT_SECRET is required'),
-  NEXT_PUBLIC_ENABLE_ANALYTICS: z.enum(['true', 'false']).optional(),
-  NEXT_PUBLIC_TRACKING_ID: z.string().optional(),
-});
-
-export function validateEnv(env: NodeJS.ProcessEnv) {
-  const result = EnvSchema.safeParse(env);
-  if (!result.success) {
-    const missing = result.error.issues.map((i) => i.path.join('.')).join(', ');
-    throw new Error(`Missing required environment variables: ${missing}`);
-  }
-  return result.data;
 }
